Split out EgressNetworkPolicy-monitoring code into its own file
